multi worker dataloader working

+++ maskmm/datagen/anchors.py
@@ -40,7 +40,6 @@ def generate_anchors(scales, ratios, shape, feature_stride, anchor_stride):
                             box_centers + 0.5 * box_sizes], axis=1)
     return boxes
 
-
 def generate_pyramid_anchors(scales, ratios, feature_shapes, feature_strides,
                              anchor_stride):
     """Generate anchors at different levels of a feature pyramid. Each scale
@@ -58,4 +57,5 @@ def generate_pyramid_anchors(scales, ratios, feature_shapes, feature_strides,
     for i in range(len(scales)):
         anchors.append(generate_anchors(scales[i], ratios, feature_shapes[i],
                                         feature_strides[i], anchor_stride))
-    return np.concatenate(anchors, axis=0)
+    # float to enable calcs wth bboxes which need to use log function
+    return np.concatenate(anchors, axis=0).astype(np.float32)
